+++ examples/with-supabase/pages/index.tsx
@@ -1,29 +1,28 @@
 import React, { useState, useEffect } from "react";
 import Head from "next/head";
 import styles from "../styles/Home.module.css";
-import ThirdPartyEmailPassword from "supertokens-auth-react/recipe/thirdpartyemailpassword";
+import ThirdPartyEmailPassword, {
+    ThirdPartyEmailPasswordAuth,
+} from "supertokens-auth-react/recipe/thirdpartyemailpassword";
 import dynamic from "next/dynamic";
 import { useSessionContext } from "supertokens-auth-react/recipe/session";
 import { getSupabase } from "../utils/supabase";
 
-const ThirdPartyEmailPasswordAuthNoSSR = dynamic(
-    new Promise<typeof ThirdPartyEmailPassword.ThirdPartyEmailPasswordAuth>((res) =>
-        res(ThirdPartyEmailPassword.ThirdPartyEmailPasswordAuth)
-    ),
-    { ssr: false }
-);
-
 export default function Home() {
     return (
-        <ThirdPartyEmailPasswordAuthNoSSR>
+        <ThirdPartyEmailPasswordAuth>
             <ProtectedPage />
-        </ThirdPartyEmailPasswordAuthNoSSR>
+        </ThirdPartyEmailPasswordAuth>
     );
 }
 
 function ProtectedPage() {
     // retrieve the authenticated user's accessTokenPayload and userId from the sessionContext
     const sessionContext = useSessionContext();
+
+    if (sessionContext.loading === true) {
+        return null;
+    }
 
     const [userEmail, setEmail] = useState("");
     useEffect(() => {