+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+-   Updates supabase example app to use supertokens-auth-react version 0.24
+
 ## [0.24.1] - 2022-07-12
 
 ### Added